requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
httpx==0.27.2
PyMuPDF==1.24.14
python-docx==1.1.2
google-generativeai==0.8.3
pydantic==2.10.3
opencv-python-headless==4.10.0.84
Pillow==11.0.0
gTTS==2.5.3
moviepy==1.0.3
